chore: Adding latestLog.json to git ignore and creating an example file to help
The latestLog file is gonna change a lot (and it varies from PC to PC, so I decided
to not sync it to the git, since it wouldn't make sense. I left a logExample.json file
so you can have a hint on how the logs are gonna look like once the application is
up and running in your equipment. Also renamed a .html file inside resources folder.
That .html is a template I use (literally a .html extraction I made of my sms page
within my home network, and put it on a localhost web server to test it out without
messing around with my legit info) - It might help you if you know how to stablish
a localhost server. That way, you could test it out even if you don't have a sms
Net Adapter II in hands, and see how everything would work out. And of course, you
can also change some flags (more in depth details on the readme.md file of this repo)
so you can force a specific situation to test the behaviour of the app, see if it's
correctly dispatching the email or not, etc.