chore: Deleting output/latestLog.json
As explained on the last commit, latestLog will change all the time, and it's personal from people to people, so it doesn't make sense to leave it to the public. I decided to left an examplee file inside 'output' folder, so anyone can have an idea on how their "latestLog.json" file will look like.