feat: Remember when I said it wasn't going to be many more feat commits? I lied
So, many things were done, I added a check for anti-flood measures, so when an email is sent to report outages in the UPS, it'll only send another email, after 30 minutes or more has passed from this last email, unless theres different status gathered by the agent, for example: If in this last email, the test flag was turned off, but in the next check, within 10 minutes it's turned on, then, the email will be sent, if not, it'll wait for 30 minutes (or for the next different flag to be captured by the agent) to dispatch the next email.

Also added an intelligence to acknowledge whether the previous check gathered errors or not, that way, in case of it gathering successes in this current check, it'll also dispatch an email, reporting the UPS has recovered itself from an outage! That is the only success scenario where an email will be sent. If 2 successes are consecutive for example, it'll not send emails, as usual, so the condition is: 1 error log -> if the next is a success, an email reporting the recovery will be sent, if not, it'll check if theres different error flags or 30 minutes has passed, to then send another error report email.

I made the html dynamic, so the email can differ what's green flags or red flags (successes or errors respectively), done a lot of testing, so it seems to be working just fine.

Also renamed a lot of files for better organization and follow good patterns, and move the manager .js's to a 'wraps' folder, again, for a better organization and pattern.

Updated the .gitignore file to exclude personalLogging files that can differ from one setup to another, so aren't absolutely necessary files to have. Kept the 'logExample.json' for the purpose of illustrating how you should expect the info to be saved in your setup.

I will still update the README a lot, so I can explain how everything must be set up. So expect a lot of commits on it. For now, that's it. It's been a helluva ride!