chore: Added a script to automatically run the node agent in windows
I made a .bat script to automatically run the node code and monitor the UPS. To make it work,
you'll need (if using windows) to add it on task scheduler, with few configurations, which I'm
gonna link in the README.md file once I'm done. But to put it in a nutshell, you'll need to
setup up 2 triggers, one in logon (and then marking "activate" and setting up any date), and
another with the option "in a schedule", where you set a start date, and then you'll mark the
option "repeat each", and select 10 minutes. That way, the agent will check every 10 minutes
for new status of the UPS, and deliver it as usual, in the log files, and the operational log
files the .bat script will generate, which goes in a folder named "operational_logs", inside
this application (you won't see it on the repo because I added to .gitignore). At this point
I'm not thinking on linux scripts, since my server is a Pi4 running windows, but it should
be very similar to the .bat script, in terms of logic, the sintax will obviously change, but
you can do the trick, specially with helpul stuff like GPT or stack overflow. Enjoy!

**OBS: I recommend (as I'll mention in the readme of the repo) to always move the application
to your "C:\node" folder, that way you won't have to alter any paths. If you don't want to use
this app inside that folder, remember you'll have to change the absolute paths inside the .bat
script, so it can work properly with your chosen path. That's it for now!