imp: Finished the changes on the readme + changed some improvements on the flow
It's been 2 days since I finished the integration and started to run it on my server,
and I had some trouble with the .cmd scripts that I was running in the automation
before. After some testing, I decided to go with a simplier way, by using a one-line
powershell execution that did the very same thing as the previous CMD script. Also
decided to split to actions of automating the agent execution and cleaning obsolete
log files in 2 distinct routines. That way it would ease the process consuming of the
server (which is very very slow), and also give me easear maintainance for future
occasions. After a lot of testing, it seems to be working flawlessly.

I also conigured a timeout for the tasks execution. I realized that once a writting
process was running, competing with the execution of the information gathering of the
agent, tha could make it loop and freeze out on screen for several minutes or even
hours, ruining the whole automation. To prevent that, since the system is very slow,
I applied a maximum execution time of 8 minutes. I realized on the testing, that even
though the raspberry is very slow running windows10, the agent execution shouldn't take
more than 2 minutes, considering scenaris where it would send emails or not. And since I
set the Task Scheduler to run it every 10 minutes, in rounded times (09:00, 09:10, 09:20
and on and on), I pictured 8 minutes to be a good window. That way, if the automation
froze (in minute of a an hour for example), in minute 8 it would be closed out, and on
the next tick (minute 10 of that hour), it would try again to run the agent. For all the
tests I ran, that really did the trick.

Beyond that, I added a few 'console.log' instructions on the code, to actually log on the
operational log .txt file the result of that execution, so it would make clearer to the
user what actually happen on that day and time. Before, it wasn't logging anything in case
of errors, or success after an error. So the user wouldn't see any message informing the UPS
was down, or that was up again after a outage. Now there is a log for each scenario. The rest
remains the same. For now, that was the major part of the job I still had to do. Future
commits will probably be mostly for minor fixes or improvements. It's been a ride so far!